5.Building_Rest_APIs_with_FastAPI/assignment2/assignment2_requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0
aiosmtplib==2.0.1
jinja2==3.1.2
google-auth-oauthlib==1.1.0
google-auth==2.25.2
requests==2.31.0
PyJWT==2.10.1  
